+++ node/index.ts
@@ -1,11 +1,17 @@
 import { Service } from '@vtex/api'
 
 import { Clients } from './clients'
+import { renderPage } from './middlewares/renderPage'
+
+const middlewares = [renderPage]
 
 export default new Service({
   clients: {
     implementation: Clients,
     options: {},
   },
-  routes: {},
+  routes: {
+    checkout: middlewares,
+    cart: middlewares,
+  },
 })